feat: restrict database initialization actions to admin users

+++ src/dev.php
@@ -9,42 +9,43 @@
 
 <body>
     <div>
-        <h2>Init</h2>
-        <form method="post">
-            <input type="submit" name="init" value="Init Database">
-        </form>
-        <?php
-        if (isset($_POST['init'])) {
-            include 'db/initDb.php';
-            initDatabase();
-            echo "database tables initialized";
-        }
-        ?>
-        <h2>Delete</h2>
-        <form method="post">
-            <input type="submit" name="delete" value="Delete Database">
-        </form>
-        <?php
-        if (isset($_POST['delete'])) {
-            include 'db/deleteDb.php';
-            deleteDatabaseTables();
-            echo "database tables deleted";
-        }
-        ?>
-        <h2>Init Teachers</h2>
-        <form method="post">
-            <input type="submit" name="init_teachers" value="Init Teachers">
-        </form>
-        <?php
-        if (isset($_POST['init_teachers'])) {
-            include 'db/initDb.php';
-            if (initTeacherInformation()) {
-                echo "teachers initialized";
-            } else {
-                echo "teachers already initialized";
+        <?php if ($_COOKIE['admin_password'] === $_ENV['ADMIN_PASSWORD']) { ?>
+            <h2>Init</h2>
+            <form method="post">
+                <input type="submit" name="init" value="Init Database">
+            </form>
+            <?php
+            if (isset($_POST['init'])) {
+                include 'db/initDb.php';
+                initDatabase();
+                echo "database tables initialized";
             }
-        }
-        ?>
+            ?>
+            <h2>Delete</h2>
+            <form method="post">
+                <input type="submit" name="delete" value="Delete Database">
+            </form>
+            <?php
+            if (isset($_POST['delete'])) {
+                include 'db/deleteDb.php';
+                deleteDatabaseTables();
+                echo "database tables deleted";
+            }
+            ?>
+            <h2>Init Teachers</h2>
+            <form method="post">
+                <input type="submit" name="init_teachers" value="Init Teachers">
+            </form>
+            <?php
+            if (isset($_POST['init_teachers'])) {
+                include 'db/initDb.php';
+                if (initTeacherInformation()) {
+                    echo "teachers initialized";
+                } else {
+                    echo "teachers already initialized";
+                }
+            } ?>
+        <?php } ?>
     </div>
 </body>
